refactor reporting batch status route
