+++ src/app/router_registry.py
@@ -219,6 +219,7 @@ from app.routers.reporting import router as reporting_router
 from app.routers.reporting_batch_controls import (
     controls_router as reporting_batch_controls_router,
 )
+from app.routers.reporting_batch_status import status_router as reporting_batch_status_router
 from app.routers.reporting_batch_worker import worker_router as reporting_batch_worker_router
 from app.routers.reporting_batches import batches_router as reporting_batches_router
 from app.routers.reporting_job_controls import controls_router as reporting_job_controls_router
@@ -397,6 +398,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(reporting_jobs_router)
     app.include_router(reporting_job_controls_router)
     app.include_router(reporting_batches_router)
+    app.include_router(reporting_batch_status_router)
     app.include_router(reporting_batch_controls_router)
     app.include_router(reporting_batch_worker_router)
     app.include_router(reporting_schedules_router)
